Restructure the code a little
